requirements.txt
asgiref==3.5.0
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
python-dotenv==0.20.0
sqlparse==0.4.2
tzdata==2022.1
